Revert "update PageWrapper with banner"
This reverts commit ef032462e4587aeb403c020c5dcd8b6dc1d65f06.
