+++ src/components/styles/PageWrapper.tsx
@@ -1,0 +1,7 @@
+import styled from "@emotion/styled/macro";
+
+export const PageWrapper = styled.div`
+  padding: 0 8px;
+  padding-bottom: 56px;
+  background-color: white;
+`;
