+++ src/components/styles/index.ts
@@ -6,6 +6,7 @@ import {
   PageTitle,
   PageButtonRow,
 } from "./Layout";
+import { PageWrapper } from "./PageWrapper";
 import { BoldStyledLink, StyledLink, StyledRouterLink } from "./StyledLink";
 import { SiderCard } from "./SiderCard";
 import { FiltersWrapper, StyledInput } from "./filters";
@@ -15,6 +16,7 @@ export {
   PageContent,
   PageLayout,
   PageSider,
+  PageWrapper,
   StyledLink,
   BoldStyledLink,
   SiderCard,
