+++ src/pages/Patch.tsx
@@ -2,8 +2,12 @@ import React, { useEffect } from "react";
 import { useParams } from "react-router-dom";
 import { BreadCrumb } from "components/Breadcrumb";
 import { PageTitle } from "components/PageTitle";
-import { PageWrapper } from "components/PageWrapper";
-import { PageContent, PageLayout, PageSider } from "components/styles";
+import {
+  PageWrapper,
+  PageContent,
+  PageLayout,
+  PageSider,
+} from "components/styles";
 import { useQuery } from "@apollo/react-hooks";
 import { GET_PATCH } from "gql/queries/patch";
 import { PatchQuery, PatchQueryVariables } from "gql/generated/types";
@@ -13,12 +17,18 @@ import get from "lodash/get";
 import { Metadata } from "pages/patch/Metadata";
 import { useHistory } from "react-router-dom";
 import { paths } from "constants/routes";
-import { useBannerDispatchContext } from "context/banners";
+import {
+  useBannerDispatchContext,
+  useBannerStateContext,
+} from "context/banners";
+import { Banners } from "components/Banners";
 import { PatchStatusBadge } from "components/PatchStatusBadge";
+import { withBannersContext } from "hoc/withBannersContext";
 
-export const Patch: React.FC = () => {
+const PatchCore: React.FC = () => {
   const { id } = useParams<{ id: string }>();
   const dispatchBanner = useBannerDispatchContext();
+  const bannersState = useBannerStateContext();
   const router = useHistory();
   const { data, loading, error, stopPolling } = useQuery<
     PatchQuery,
@@ -40,10 +50,15 @@ export const Patch: React.FC = () => {
     router.push(`${paths.patch}/${id}/configure`);
   }
   if (error) {
-    return <PageWrapper />;
+    return (
+      <PageWrapper>
+        <Banners banners={bannersState} removeBanner={dispatchBanner.remove} />
+      </PageWrapper>
+    );
   }
   return (
     <PageWrapper data-cy="patch-page">
+      <Banners banners={bannersState} removeBanner={dispatchBanner.remove} />
       {patch && <BreadCrumb patchNumber={patch.patchNumber} />}
       <PageTitle
         loading={loading}
@@ -65,3 +80,5 @@ export const Patch: React.FC = () => {
     </PageWrapper>
   );
 };
+
+export const Patch = withBannersContext(PatchCore);
